github-action-secrets/requirements.txt
pylint>=2.4.4
PyNaCl>=1.3.0
requests>=2.22.0
click
